+++ Orders/BackEnd/Services/SubElementService.cs
@@ -1,6 +1,8 @@
-﻿using BackEnd.Models;
+﻿using BackEnd.ErrorResponseModel;
+using BackEnd.Models;
 using BackEnd.Models.Dto.SubElement;
 using BackEnd.Repository;
+using System.Net;
 
 namespace BackEnd.Services
 {
@@ -15,6 +17,27 @@ namespace BackEnd.Services
 
         public async Task CreateSubElement(SubElementDto subElementDto)
         {
+            if (string.IsNullOrEmpty(subElementDto.Type))
+            {
+                throw new HttpStatusException(HttpStatusCode.BadRequest, "Type cannot be empty");
+            }
+            else if (subElementDto.Width == null)
+            {
+                throw new HttpStatusException(HttpStatusCode.BadRequest, "Width cannot be empty");
+            }
+            else if (subElementDto.Height == null)
+            {
+                throw new HttpStatusException(HttpStatusCode.BadRequest, "Height cannot be empty");
+            }
+            else if (subElementDto.Element == null)
+            {
+                throw new HttpStatusException(HttpStatusCode.BadRequest, "Element cannot be empty");
+            }
+            else if (subElementDto.WindowId == null || subElementDto.WindowId == 0)
+            {
+                throw new HttpStatusException(HttpStatusCode.BadRequest, "Window id cannot be empty");
+            }
+
             var subElement = new SubElement()
             {
                 Type = subElementDto.Type,
@@ -28,12 +51,23 @@ namespace BackEnd.Services
 
         public async Task DeleteSubElement(int subElementId)
         {
-            await _SubElementRepository.DeleteSubElement(subElementId);
+            if(subElementId == null && subElementId == 0)
+            {
+                throw new HttpStatusException(HttpStatusCode.BadRequest, "Id cannot be null or zero");
+                
+            }
+                await _SubElementRepository.DeleteSubElement(subElementId);
         }
 
         public async Task<SubElement> GetSubElement(int subElementId)
         {
-            return await _SubElementRepository.GetSubElement(subElementId);
+            if (subElementId == null && subElementId == 0)
+            {
+                throw new HttpStatusException(HttpStatusCode.BadRequest, "Id cannot be null or zero");
+                
+            }
+                return await _SubElementRepository.GetSubElement(subElementId);
+
         }
 
         public async Task<IList<SubElement>> GetSubElements()
