+++ Orders/BackEnd/Services/WindowService.cs
@@ -1,8 +1,10 @@
-﻿using BackEnd.Models;
+﻿using BackEnd.ErrorResponseModel;
+using BackEnd.Models;
 using BackEnd.Models.Dto;
 using BackEnd.Models.Dto.SubElement;
 using BackEnd.Models.Dto.Window;
 using BackEnd.Repository;
+using System.Net;
 
 namespace BackEnd.Services
 {
@@ -17,6 +19,18 @@ namespace BackEnd.Services
 
         public async Task CreateWindow(WindowDto windowDto)
         {
+            if(string.IsNullOrEmpty(windowDto.Name))
+            {
+                throw new HttpStatusException(HttpStatusCode.BadRequest, "Name cannot be empty");
+            }
+            if(windowDto.QuantityOfWindows == null)
+            {
+                throw new HttpStatusException(HttpStatusCode.BadRequest, "Quantity of windows cannot be empty");
+            }
+            else if (windowDto.OrderId == null || windowDto.OrderId == 0)
+            {
+                throw new HttpStatusException(HttpStatusCode.BadRequest, "Order id cannot be empty");
+            }
             var window = new Window()
             {
                 Name = windowDto.Name,
@@ -28,11 +42,20 @@ namespace BackEnd.Services
 
         public async Task DeleteWindow(int windowId)
         {
-            await _windowRepository.DeleteWindow(windowId);
+            if(windowId == 0 && windowId == null)
+            {
+                throw new HttpStatusException(HttpStatusCode.BadRequest, "Id cannot be null or zero");
+            }
+                await _windowRepository.DeleteWindow(windowId);
         }
 
         public async Task<GetWindowDto> GetWindow(int windowId)
         {
+            if (windowId == 0 || windowId == null)
+            {
+                throw new HttpStatusException(HttpStatusCode.BadRequest, "Id cannot be null or zero");
+            }
+           
             var window = await _windowRepository.GetWindow(windowId);
 
             var getWindowDto = new GetWindowDto()
